tflite-cpp/test-model-generator/pyproject.toml
[tool.poetry]
name = "test-model-generator"
version = "0.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8,<3.12"
tensorflow-macos = { version = "2.12.0", platform = "darwin" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
